uncompressed: pulled RowInterleaveDecoder into its own file
